# frontend/types/api/openapi.json
{
  "components": {
    "schemas": {
      "AISuggestionHistoryItem": {
        "description": "One entry in the proposal history for a single (instance, field) coord.\n\nSame shape as AISuggestionItem minus `status` \u2014 history is proposal trail only.",
        "properties": {
          "confidence_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "field_id": {
            "format": "uuid",
            "title": "Field Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "instance_id": {
            "format": "uuid",
            "title": "Instance Id",
            "type": "string"
          },
          "proposed_value": {
            "additionalProperties": true,
            "title": "Proposed Value",
            "type": "object"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "instance_id",
          "field_id",
          "proposed_value",
          "confidence_score",
          "rationale",
          "created_at",
          "evidence"
        ],
        "title": "AISuggestionHistoryItem",
        "type": "object"
      },
      "AISuggestionItem": {
        "description": "A single AI suggestion (latest proposal per coord) with caller-scoped status.",
        "properties": {
          "confidence_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "field_id": {
            "format": "uuid",
            "title": "Field Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "instance_id": {
            "format": "uuid",
            "title": "Instance Id",
            "type": "string"
          },
          "proposed_value": {
            "additionalProperties": true,
            "title": "Proposed Value",
            "type": "object"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "instance_id",
          "field_id",
          "proposed_value",
          "confidence_score",
          "rationale",
          "created_at",
          "evidence",
          "status"
        ],
        "title": "AISuggestionItem",
        "type": "object"
      },
      "AISuggestionsResponse": {
        "description": "Response for the suggestions endpoint.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/AISuggestionItem"
            },
            "title": "Suggestions",
            "type": "array"
          }
        },
        "required": [
          "suggestions",
          "count"
        ],
        "title": "AISuggestionsResponse",
        "type": "object"
      },
...
      "ApiResponse_AISuggestionsResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AISuggestionsResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dados da resposta"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ],
            "description": "Error details"
          },
          "ok": {
            "description": "Indica se a operacao foi bem-sucedida",
            "title": "Ok",
            "type": "boolean"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "rastreamento",
            "title": "Trace Id"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiResponse[AISuggestionsResponse]",
        "type": "object"
      },
...
      "ApiResponse_Union_RunSummaryResponse__NoneType__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunSummaryResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dados da resposta"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ],
            "description": "Error details"
          },
          "ok": {
            "description": "Indica se a operacao foi bem-sucedida",
            "title": "Ok",
            "type": "boolean"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "rastreamento",
            "title": "Trace Id"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiResponse[Union[RunSummaryResponse, NoneType]]",
        "type": "object"
      },
...
      "ApiResponse_list_AISuggestionHistoryItem__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AISuggestionHistoryItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dados da resposta",
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ],
            "description": "Error details"
          },
          "ok": {
            "description": "Indica se a operacao foi bem-sucedida",
            "title": "Ok",
            "type": "boolean"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "rastreamento",
            "title": "Trace Id"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiResponse[list[AISuggestionHistoryItem]]",
        "type": "object"
      },
...
      "ApiResponse_list_ArticleRunRef__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ArticleRunRef"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dados da resposta",
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ],
            "description": "Error details"
          },
          "ok": {
            "description": "Indica se a operacao foi bem-sucedida",
            "title": "Ok",
            "type": "boolean"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "rastreamento",
            "title": "Trace Id"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiResponse[list[ArticleRunRef]]",
        "type": "object"
      },
      "ApiResponse_list_UUID__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dados da resposta",
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ],
            "description": "Error details"
          },
          "ok": {
            "description": "Indica se a operacao foi bem-sucedida",
            "title": "Ok",
            "type": "boolean"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "rastreamento",
            "title": "Trace Id"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiResponse[list[UUID]]",
        "type": "object"
      },
...
      "ArticleRunRef": {
        "description": "Per-article run reference returned by POST /articles/form-runs.\n\n``run_id`` is None when the article has no matching run.",
        "properties": {
          "article_id": {
            "format": "uuid",
            "title": "Article Id",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "required": [
          "article_id",
          "run_id"
        ],
        "title": "ArticleRunRef",
        "type": "object"
      },
...
      "ErrorDetail": {
        "description": "Standardized error details.",
        "properties": {
          "code": {
            "description": "Error code (see ApiErrorCode)",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Additional error details",
            "title": "Details"
          },
          "message": {
            "description": "Error message",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "FormRunsRequest": {
        "description": "Request body for POST /api/v1/articles/form-runs.\n\nResolves the latest relevant run for each article_id in the batch.\n``project_id`` is used for BOLA enforcement.",
        "properties": {
          "article_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Article Ids",
            "type": "array"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "template_id": {
            "format": "uuid",
            "title": "Template Id",
            "type": "string"
          }
        },
        "required": [
          "article_ids",
          "template_id",
          "project_id"
        ],
        "title": "FormRunsRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunSummaryResponse": {
        "properties": {
          "article_id": {
            "format": "uuid",
            "title": "Article Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "format": "uuid",
            "title": "Created By",
            "type": "string"
          },
          "hitl_config_snapshot": {
            "additionalProperties": true,
            "title": "Hitl Config Snapshot",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "results": {
            "additionalProperties": true,
            "title": "Results",
            "type": "object"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "template_id": {
            "format": "uuid",
            "title": "Template Id",
            "type": "string"
          },
          "version_id": {
            "format": "uuid",
            "title": "Version Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "article_id",
          "template_id",
          "kind",
          "version_id",
          "stage",
          "status",
          "hitl_config_snapshot",
          "parameters",
          "results",
          "created_at",
          "created_by"
        ],
        "title": "RunSummaryResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Supabase JWT": {
        "description": "JWT token do Supabase Auth",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend API for Prumo - Systematic Review Platform",
    "title": "Prumo API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/articles/form-runs": {
      "post": {
        "description": "Resolve the latest relevant run per article for the extraction form.\n\nPer article: returns the latest non-terminal run; falls back to the\nlatest finalized run; returns run_id=null when no run exists.\nCancelled runs are excluded. BOLA-gated via project_id in the body.",
        "operationId": "post_form_runs_api_v1_articles_form_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormRunsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_ArticleRunRef__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Supabase JWT": []
          }
        ],
        "summary": "Post Form Runs",
        "tags": [
          "articles"
        ]
      }
    },
    "/api/v1/articles/{article_id}/active-run": {
      "get": {
        "description": "Return the latest non-terminal extraction run for the article.\n\nReturns null (data: null) when no active run exists. 404 when the\narticle is not found; 403 when the caller is not a project member.",
        "operationId": "get_active_run_api_v1_articles__article_id__active_run_get",
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Article Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "template_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Union_RunSummaryResponse__NoneType__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Supabase JWT": []
          }
        ],
        "summary": "Get Active Run",
        "tags": [
          "articles"
        ]
      }
    },
...
    "/api/v1/articles/{article_id}/finalized-run": {
      "get": {
        "description": "Return the latest finalized extraction run for the article.\n\nReturns null (data: null) when no finalized run exists. 404 when the\narticle is not found; 403 when the caller is not a project member.",
        "operationId": "get_finalized_run_api_v1_articles__article_id__finalized_run_get",
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Article Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "template_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Union_RunSummaryResponse__NoneType__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Supabase JWT": []
          }
        ],
        "summary": "Get Finalized Run",
        "tags": [
          "articles"
        ]
      }
    },
    "/api/v1/articles/{article_id}/instance-ids": {
      "get": {
        "description": "Return all extraction_instance ids for the article.\n\n404 when article is not found; 403 when caller is not a project member.",
        "operationId": "get_instance_ids_api_v1_articles__article_id__instance_ids_get",
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Article Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_UUID__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Supabase JWT": []
          }
        ],
        "summary": "Get Instance Ids",
        "tags": [
          "articles"
        ]
      }
    },
    "/api/v1/articles/{article_id}/suggestions": {
      "get": {
        "description": "Return latest AI proposals per (instance, field) with caller-scoped status.\n\nStatus is derived exclusively from the caller's own reviewer_state rows \u2014\nnever another reviewer's (blind boundary, Constraint 3).\n404 when article is not found; 403 when caller is not a project member.",
        "operationId": "get_suggestions_api_v1_articles__article_id__suggestions_get",
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Article Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "instance_ids",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "format": "uuid",
                "type": "string"
              },
              "title": "Instance Ids",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AISuggestionsResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Supabase JWT": []
          }
        ],
        "summary": "Get Suggestions",
        "tags": [
          "articles"
        ]
      }
    },
    "/api/v1/articles/{article_id}/suggestions/history": {
      "get": {
        "description": "Return AI proposal history for a single (instance, field) coord.\n\nNewest-first; no status (history is the raw proposal trail).\n404 when article is not found; 403 when caller is not a project member.",
        "operationId": "get_suggestions_history_api_v1_articles__article_id__suggestions_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Article Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "instance_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Instance Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "field_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Field Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_AISuggestionHistoryItem__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Supabase JWT": []
          }
        ],
        "summary": "Get Suggestions History",
        "tags": [
          "articles"
        ]
      }
    },
...
  }
}